+++ 05-SceneManager/PlayScene.cpp
@@ -84,7 +84,7 @@ void CPlayScene::_ParseSection_ANIMATIONS(string line)
 	for (int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
 	{
 		int sprite_id = atoi(tokens[i].c_str());
-		int frame_time = atoi(tokens[i+1].c_str());
+		int frame_time = atoi(tokens[static_cast<__int64>(i)+1].c_str());
 		ani->Add(sprite_id, frame_time);
 	}
 
@@ -485,7 +485,7 @@ void CPlayScene::Update(DWORD dt)
 	{
 		if(cx>=MAP2_WIDTH- game->GetBackBufferWidth())
 		{
-			cx = MAP2_WIDTH - game->GetBackBufferWidth();
+			cx = MAP2_WIDTH - float(game->GetBackBufferWidth());
 		}
 		cy = 0;
 	}
@@ -493,10 +493,10 @@ void CPlayScene::Update(DWORD dt)
 	{
 		if (cx >= MAP1_WIDTH - game->GetBackBufferWidth())
 		{
-			cx = MAP1_WIDTH - game->GetBackBufferWidth();
+			cx = MAP1_WIDTH - float(game->GetBackBufferWidth());
 		}
 		if (cy >= MAP1_HEIGHT - game->GetBackBufferHeight() + HUD_HEIGHT) {
-			cy = MAP1_HEIGHT - game->GetBackBufferHeight() + HUD_HEIGHT;
+			cy = MAP1_HEIGHT -float( game->GetBackBufferHeight()) + HUD_HEIGHT;
 		}
 	}
